requirements.txt
MouseInfo==0.1.3
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pywin32==301
